Fix 500+ build errors across Engine project
Major fixes include:
- Archetype comparisons: Changed string comparisons to CharacterClass enum
- Result factory methods: Changed Success() to Successful() across 18+ files
- ILogger methods: Fixed Debug() to LogDebug(), BeginOperation() to BeginScope()
- Type conversions: Fixed GridPosition (Zone/Row/Column), EquipmentType, nullables
- HazardTrigger enum: Unified Engine and Core versions
- Direction enum: Added using aliases to resolve namespace conflicts
- Service constructors: Fixed missing parameters for multiple services
- EnemyAction: Fixed namespace conflicts between Core and Engine versions
- Repository methods: Fixed DescriptorRepository method names
- Added missing properties to PlayerStrategy, RotationStep, GridTile classes
- Fixed DungeonGenerator LootNode abstract class usage
- Fixed CombatEngine undeclared variables and type mismatches

Files modified: 65 across Core, Engine, and Tests projects

+++ RuneAndRust.Engine/EndlessMode/EndlessModeService.cs
@@ -418,8 +418,8 @@ public class EndlessModeService
             TotalWaves = waves.Count,
             BossWavesCompleted = waves.Count(w => w.IsBossWave && w.IsComplete),
             AverageWaveTime = waves.Any(w => w.WaveTimeSeconds.HasValue)
-                ? waves.Where(w => w.WaveTimeSeconds.HasValue).Average(w => w.WaveTimeSeconds!.Value)
-                : 0,
+                ? (float)waves.Where(w => w.WaveTimeSeconds.HasValue).Average(w => w.WaveTimeSeconds!.Value)
+                : 0f,
             FastestWave = waves.Where(w => w.WaveTimeSeconds.HasValue).MinBy(w => w.WaveTimeSeconds!.Value),
             SlowestWave = waves.Where(w => w.WaveTimeSeconds.HasValue).MaxBy(w => w.WaveTimeSeconds!.Value)
         };
